+++ exonum-java-binding/common/src/main/java/com/exonum/binding/common/proofs/list/ListProofVisitor.java
@@ -23,4 +23,6 @@ public interface ListProofVisitor {
   void visit(ListProofHashNode listProofHashNode);
 
   void visit(ListProofElement value);
+
+  void visit(ListProofOfAbsence listProofOfAbsence);
 }