+++ exonum-java-binding/common/src/test/java/com/exonum/binding/common/proofs/map/UncheckedFlatMapProofTest.java
@@ -18,6 +18,9 @@ package com.exonum.binding.common.proofs.map;
 
 import static com.exonum.binding.common.hash.Funnels.hashCodeFunnel;
 import static com.exonum.binding.common.proofs.DbKeyFunnel.dbKeyFunnel;
+import static com.exonum.binding.common.proofs.map.UncheckedFlatMapProof.BLOB_PREFIX;
+import static com.exonum.binding.common.proofs.map.UncheckedFlatMapProof.MAP_NODE_PREFIX;
+import static com.exonum.binding.common.proofs.map.UncheckedFlatMapProof.MAP_ROOT_PREFIX;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singleton;
 import static java.util.Collections.singletonList;
@@ -109,10 +112,20 @@ class UncheckedFlatMapProofTest {
     ByteString value = FIRST_VALUE;
     MapEntry<ByteString, ByteString> mapEntry = createMapEntry(key, value);
 
-    HashCode valueHash = HASH_FUNCTION.hashByteString(value);
-    HashCode expectedRootHash = HASH_FUNCTION.newHasher()
+    HashCode valueHash = HASH_FUNCTION.newHasher()
+        .putByte(BLOB_PREFIX)
+        .putBytes(value.toByteArray())
+        .hash();
+
+    HashCode singleEntryHash = HASH_FUNCTION.newHasher()
+        .putByte(MAP_NODE_PREFIX)
         .putObject(DbKey.newLeafKey(key), dbKeyFunnel())
         .putObject(valueHash, hashCodeFunnel())
+        .hash();
+
+    HashCode expectedRootHash = HASH_FUNCTION.newHasher()
+        .putByte(MAP_ROOT_PREFIX)
+        .putObject(singleEntryHash, hashCodeFunnel())
         .hash();
 
     UncheckedMapProof uncheckedFlatMapProof =
@@ -122,7 +135,7 @@ class UncheckedFlatMapProofTest {
             emptyList());
     CheckedMapProof checkedMapProof = uncheckedFlatMapProof.check();
 
-    assertThat(checkedMapProof.getRootHash(), equalTo(expectedRootHash));
+    assertThat(checkedMapProof.getIndexHash(), equalTo(expectedRootHash));
 
     assertThat(checkedMapProof.getEntries(), equalTo(singleton(mapEntry)));
     assertTrue(checkedMapProof.containsKey(key));
@@ -149,8 +162,14 @@ class UncheckedFlatMapProofTest {
     UncheckedMapProof uncheckedFlatMapProof =
         new UncheckedFlatMapProof(emptyList(), emptyList(), emptyList());
 
+    HashCode expectedRootHash = HASH_FUNCTION.newHasher()
+        .putByte(MAP_ROOT_PREFIX)
+        .putBytes(new byte[Hashing.DEFAULT_HASH_SIZE_BYTES])
+        .hash();
+
     CheckedMapProof checkedMapProof = uncheckedFlatMapProof.check();
     assertThat(checkedMapProof.getProofStatus(), equalTo(MapProofStatus.CORRECT));
+    assertThat(checkedMapProof.getIndexHash(), equalTo(expectedRootHash));
   }
 
   @Test