+++ exonum-java-binding/core/rust/exonum-java/start_qa_node.sh
@@ -32,7 +32,7 @@ echo "PROJ_ROOT=${EJB_ROOT}"
 
 source "${EJB_ROOT}/tests_profile"
 
-java_library_path="${EJB_ROOT}/core/rust/target/debug"
+java_library_path="${EJB_ROOT}/core/rust/target/debug/deps"
 
 # Find the artifact
 ARTIFACT_PATH="$(find ${EJB_ROOT} -type f -name exonum-java-binding-qa-service-*-artifact.jar)"
@@ -52,6 +52,9 @@ rm -rf testnet
 mkdir testnet
 
 trap "killall exonum-java" SIGINT SIGTERM EXIT
+
+# Delete the java_bindings library from the target/debug if any to prevent ambiguity in dynamic linking (ECR-3468)
+rm -f ${EJB_ROOT}/core/rust/target/debug/libjava_bindings.*
 
 # Configure and run nodes
 node_count=$1